Extract ClaimVoucherSection component
